Update index.php
Adding selection class for currently selected day with a dotted border.

Since the date is written in the lower right of the textbox, it's hard to determine which box is selected. A user might end up editing the wrong box. Adding some css will make it more obvious which date is being edited.