Have reduceNavInfos handle leaveThreadActionTypes.success directly
Instead of having a separate `goBack` handle nav changes as the result of a LEAVE_THREAD_SUCCESS, we handle the action directly in `reduceNavInfo`. To accomplish this we introduce two new helpers: `popChatScreensForThreadID` and `filterChatScreensForThreadInfos`. Our new functionality is also more robust, since we make sure that if we are deauthorized for any threads, we remove any chat screens corresponding to those threads from the stack.