Allow possibly_reset_thread_to_unread on update_focused_threads.php to set to unread
In cases where the client navigates away before getting a message, the update_focused_threads.php call should set the thread to unread. `ping.php` will do this later, but the sooner the better.

Only downside is that this might make the client appear unread before they get the new message, but that should be rare and will get resolved quickly.