Make sure we send a notif for subthread creation
Before this, we would send a notif for the corresponding thread creation message, but that doesn't get delivered to users that have visibility on the new subthread but aren't initially members.

This adds a lot of messy code that will make sure we deliver a notif for a subthread creation message if and only if the subthread is known-of, but either isn't visible (which would prevent the corresponding thread creation message from triggering a notif) or the notif recipient isn't a member of the newly created subthread.