[server] Don't delete focused row when activity status doesn't change
Right now, whenever we get an activity update, we first delete all the rows for the user/cookie, and then recreate the valid ones.

This commit changes the logic so we first update the rows that are still valid, and then delete the ones that are no longer valid.

Since clients never send activity updates unless something has changed, and only one thing can be focused at once, this should result in minimal changes.