[native] Don't call pingNow on iosNotificationOpened
`iosNotificationOpened` only gets triggered when the app is closed or backgrounded. Thus we can expect a `ping` call to fire pretty immediately anyways, and in fact trying to fire one here can prevent the `startTimeouts` one that includes the activity update from firing, meaning the activity update never gets communicated to the server.

Android already has the same behavior here - it doesn't call `pingNow` when a notif is pressed, only when it is received.

Note that we still have an issue in terms of the initial activity update from a notif press. If the notif press leads to an activity update, we should ideally include that activity update in the initial ping, so as to avoid any races.