Send deletedEntryIDs down along with deltaEntryInfos
If a client has a `RawEntryInfo` in its store that is outside the client's current `CalendarQuery` range, that `RawEntryInfo` may become stale. If the `RawEntryInfo` gets deleted, but later its `CalendarQuery` is adjusted to include it, the server will not think to include the deleted `RawEntryInfo` in the `deltaEntryInfos` payload, as it is precluded by the `NOT_DELETED` filter. However, this will lead to the stale `RawEntryInfo` not getting updated, which will only get corrected by the next state check.

To address this, this commit introduces a `deletedEntryIDs` field that is sent along with `deltaEntryInfos`. It is only populated when the `NOT_DELETED` filter is active, and includes a list of entry IDs that are within the `CalendarQuery` range. If the client has any of these entry IDs in its `EntryStore`, it will update them to be deleted.