[server] Make sure updates get delivered in non-socket scenario
If verification or account update occurred over a non-socket context, we would drop the updates for the client.

This makes sure we deliver them via broadcast (or in the state sync payload) even in a non-socket context.

Also explicitly specifies to ignore user subscription updates, as those are handled explicitly in the client.