[native] Fix ColorPicker on iOS
`setTimeout` appears to be necessary as of React Native 0.60. `InteractionManager.runAfterInteractions` appears to have been deferring the `measure` until after `ColorPickerModal`'s slide-in animation before React Native 0.60, but as of 0.60 it is broken for release builds on device. (No issue appears on dev modes, or on release builds on simulator.)

Long-term would be best to move away from using absolute positioning data, which may be possible if we move to `react-native-gesture-handler`. Not bothering with that right now.