[native] Update Android keyboard avoiding strategy
We're using translucent status/navigation bars now, which means the soft keyboard won't resize the window. Apparently `windowSoftInputMode` is just ignored. We have to handle it ourselves with `KeyboardAvoidingView` now, just like on Android.

An exception is the media gallery, which still resizes the window, and continues to require dynamic `TabBar` hiding.