[native] Define dummyNodeForTextMessageHeightMeasurement in InnerTextMessage
Now that we have `NodeTextMeasurer` instead of `TextHeightMeasurer`, we don't need to keep a `textMessageMaxWidthSelector` around - we can just render use the same styles.

We also don't need the `Markdown` component to forward props anymore, since the dummy we pass to `NodeHeightMeasurer` is now a `View` (which can take `onLayout` directly).