[web] Fix scrollToBottom behavior in ChatMessageList
Summary:
I'm not sure if something changed between testing D71 last week and today in Chrome, but I noticed that the `scrollToBottom` mechanism I had just introduced wasn't working properly. I seem to remember last week that `scrollTop` was not 0 when Chrome was scrolled to the bottom, but it is now. Maybe I was just really confused last week.

This time I made sure to test on Chrome, Firefox, and Safari. It should all be working correctly now.

Reviewers: palys-swm, KatPo

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D95