[web] Fix ChatMessageList condition for loading more messages
Summary:
At some point in the last couple years the behavior of `scrollTop` for Chrome when using `flex-direction: column-reverse` must have changed. The current logic makes the web browser load more messages for every thread until it gets to the top, which takes a lot of memory and makes the scroll window really laggy.

I tested this on Chrome, Firefox, and Safari.

Reviewers: palys-swm, KatPo

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D97