[server] Clean up relationship row logic in thread-permission-updaters
Summary:
In D21, we were supposed to catch all places where we had been checking if a role was 0, since we had added a new role -1 that had similar logic.

I noticed a place we had missed previously. I found the logic fairly hard to understand so I updated it a bit to be more clear.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D99