[server] Create last read message column
Summary: I included creating a foreign key constraint but we can decide to remove it

Test Plan:
The script was tested and assigned correct values of last message and last read message: for the threads which were unread, last read message was not updated - it's value was 0.
I tried to run this script twice: one when the columns were not present, then the second time: the content of the columns was updated.
Also, I created a subthread of a thread, which was secret and one of the users was not a part of it. Last message and last read message were updated correctly to different values: a text message for excluded user and create subthread message for included.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, KatPo, zrebcu411

Differential Revision: https://phabricator.ashoat.com/D176