[lib] Consider a null currentUserInfo to be a session downgrade
Summary:
In 8b4425d64c980079e4f58e543221797b2bde1179 I made it possible for `LOG_OUT_SUCCESS` to have a `null` `currentUserInfo`. That should still be considered a session downgrade, and we should block it if the session has changed since the log out began.

Note that an `undefined` `currentUserInfo` can come from a `SET_NEW_SESSION`, in which case it means that the user hasn't changed, just their `sessionID` has. In that case it shouldn't be treated as a session downgrade.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D181