[native] Don't override logout with session recovery
Summary:
If a request fails due to an invalid session, our session recovery logic tries to log the user back in and to retry the request. However, if the user logs out during this process, the logout can get overriden by the session recovery.

We should respect the user's wishes and log them out in this scenario.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D183