[native] Use ratchetAlongWithKeyboardHeight for footerPaddingTop in LoggedOutModal
Summary:
`footerPaddingTop` controls where the "Forgot password?" link appears in the log-in screen. There was an issue on Android where expanding elements of the on-screen keyboard would make the link move, sometimes over `LogInPanel`.

I made `ratchetAlongWithKeyboardHeight` to handle this use case. On iOS it will "ratchet" up as the keyboard height increases, but won't go back down until the keyboard height is reset to zero. On Android, it will only set the keyboard height when it first appears, and then again when it disappears.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D203