[native] Fix scroll issue in RelationshipList
Summary:
It was impossible to scroll to the bottom. You could only scroll to 24 pixels above the bottom, because there was a `marginTop` set on the `FlatList`'s `contentContainerStyle`.

I changed it to a `padding` to avoid the scroll issue, and also added some padding to the bottom for visual reasons.

Reviewers: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D213