[server] Collapse notifs for all users instead of just the first
Summary:
Notif collapsing is something we do for certain message types. An example is if a user creates a calendar entry, and then immediately edits it before anybody opens the app, we will "collapse" the second notif into the first one.

The job of `fetchCollapsableNotifs` is to fetch all existing messages that should get collapsed together with a new message. It needs to do this for all users in the thread at once.

The behavior here was broken because the result set was getting grouped by `messageID`, but then only the first row in each group was looked at. This meant that if multiple users had the same message getting collapsed, it will only register for the first user.

I actually only noticed this bug because I saw a "expected single row, but there are multiple" error in the logs. This occurred because we were parsing all of the messages for all users together, instead of separating by user.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D224