[web] Scroll filter button along with calendar
Summary:
I'm not sure if this was always this way or something changed and broke it, but right now the filter button on the Calendar tab doesn't scroll along with the rest of the view.

The issue is that the filter button is being absolutely positioned, but its most recent `relative` ancestor is outside the scrolling element. This diff fixes it by making sure there is a `relative` ancestor within the scrolling element.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D226