[server] es-module-specifier-resolution -> experimental-specifier-resolution
Summary:
This got renamed at some point in Node 14, but the old name was kept as an alias. Let's switch to the new one.

Also, I fixed some of the BSD-style arguments to Node to be consistent with BSD style (eg. `--arg=something` rather than `--arg something`). Nodemon's `--watch` argument didn't work with this style though.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D233