[server] Fix iOS in-app notifs for threads that have push notifs disabled
Summary: We were omitting the text of the notif from the payload based on the assumption that it was only needed when the banner is displayed. But I forgot to consider that in-app notifs need the text as well.

Test Plan: This one is pretty hard to test exactly because we can't deliver notifs to the simulator through APNs. I have a test `.apns` file that I use to test, but it was a full (not badge-only) notif. I've added a second `badgeonly.apns` test file.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D242