[server] Update last message and last read message in thread permissions updater
Summary: I removed unread prop from rowToSave object and added two new rows. In the db, unread column is still updated based on last message and last read message

Test Plan: I added a new user to an existing thread. The thread was marked as unread for that user. Then I removed the user from the thread which updated both columns to 0 - we were setting unread = 0 in that case. Then I added the user again and the thread was marked as unread.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D247