[server] Apply new username restrictions for new signups from new clients
Summary:
In D244 I only applied the new restrictions to clients, not to the server.

I don't want the server to apply these restrictions to old clients because the error message will be confusing. Once `verifyClientSupported` stops supporting versions below 69 (the next version) we can make these restrictions apply to everyone.

Test Plan: I tested the old/new `codeVersions` from native with usernames that fit the old requirements but not the new

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D251