[native] Buffer chat list data updates while overlay is active
Summary: We want the screen to stay frozen while an overlay is active so that the coordinates stay the same.

Test Plan: I opened the same thread on web and on native, then focused a message on native, and sent a new message on web. I confirmed that the new message didn't appear I closed the overlay.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D253