[server] Drop unread column
Test Plan: I tested delete script by running it and it dropped the column. I created a new db, ran create-db script on it and verified that the column was not created. I also tested changes in permissions updater by adding a user to a thread and removing it - it worked correctly, showing the thread with proper unread style.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D254