[native] Handle KeyboardAvoidingView edge case where onKeyboardChange fires before onLayout
Summary:
While using `RelationshipUpdateModal` on prod I noticed an issue where sometimes `KeyboardAvoidingView` wouldn't work properly and part of the modal would be covered up by the keyboard.

I was able to repro this in dev mode (but not with the debugger, for some reason) and found it was because `onKeyboardChange` can sometimes fire before `onLayout`. Not sure if this happens on Android too, but it definitely can happen on iOS.

I'm not sure why we haven't seen this before, since we have other `Modal`s that have `TextInput`s with `autoFocus={true}`, like `ThreadPickerModal` for instance.

Anyways, this diff addresses this edge case by buffering `onKeyboardChange` calls that occur before the first `onLayout`, and firing them right after that first `onLayout` gets called.

Test Plan: I was able to repro the issue on dev mode by repeatedly opening and closing `RelationshipUpdateModal`. I added some `console.log` statements on `onKeyboardChange` and `onLayout`, and confirmed that the issue was gone, even when `onKeyboardChange` fires before `onLayout`.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D256