[server] Don't require accountPassword to change thread type or parent thread
Summary: This requirement is from an old version of the app. Right now it's actually not possible to change thread type or parent thread anywhere in the app, so it hasn't mattered. I'm introducing the ability to promote sidebars to subthreads, so I need to make sure we can hit this API without requiring the account password.

Test Plan: Made sure later diff that calls this end point to change the `type` of the thread works without `accountPassword`

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D273