[server] Avoid exception when sending message to thread with no members
Summary:
It's theoretically possible for somebody to send a message to a thread with zero members. (The sender would have to be a parent admin.)

In this case we'll get a SQL query with a `WHERE ()` clause, which will cause an error. Let's avoid running this query in this edge case.

Test Plan: I tested an unrelated bug that was tripping the same zero-members edge case

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D278