[server] Run commitMembershipChangeset before createMessages
Summary: It's important that we have the `memberships` table up-to-date before we run `createMessages`. Otherwise, for instance, push notifs about thread creation might not get delivered, or the `last_message` column might not be updated.

Test Plan: I tested subthread creation, which previously was detecting zero members. I made sure it had a nonzero value this time with a `console.log` statement

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D279