[native] Fix bug in ThreadSettings for threads with no subthreads
Summary: I accidentally missed this bug. In the process I compared the subthreads code with the members code and decided to make the members code look more like the subthreads code. (Except for the bug, which existed only in the subthreads code.)

Test Plan: I tested `ThreadSettings` on a thread without subthreads

Reviewers: palys-swm, KatPo

Reviewed By: palys-swm, KatPo

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D283