[native] Only hide MultimediaMessageMultimedia that is being shown
Summary: `MultimediaMessageMultimedia` is the component that handles displaying one of the multimedia in a `MultimediaMessage`. Right now when one `MultimediaModal` is being shown, we hide all of the `MultimediaMessageMultimedia`. But we only need to hide the one that matches the `MultimediaModal`.

Test Plan: Open up a chat with multiple multimedia visible on the same screen and make sure only the one that get pressed gets hidden

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D293