[native] Create selector which selects existing thread or a personal thread, if it exists or a thread from the params
Summary: The selector should return the thread with matching id. If there is no such thread, we try to search for the personal thread. If it does not exist, we return the pending thread.

Test Plan: Currently personal threads can't be created. I verified that for the existing threads, returned thread info is the same as it was. For the pending thread I used console log to check if it was passed to chat input bar.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D297