[native] Disable GestureTouchableOpacity while keyboardActive
Summary:
Before D164, because `ChatList` uses `FlatList`'s default value for `keyboardShouldPersistTaps`, an active keyboard would prevent the `TouchableOpacity` wrapping multimedia and text blobs in the chat view from firing.

But `GestureTouchableOpacity` works directly through `react-native-gesture-handler`, and for that reason we must handle the keyboard ourselves.

Test Plan: The issue was that the tooltips could activate while the keyboard was open, but for some reason, I could only repro it on iOS release builds. I tested it and made sure this diff fixed the issue on an iOS release build.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D300