[server] Fix Android in-app notifs for threads that have push notifs disabled
Summary: This is the Android equivalent of D242. For some reason I thought Android was working correctly when I fixed that iOS bug, but turns out it had the same issue.

Test Plan: I deployed a v68 and v69 client, one on a physical device and one on an emulator. I then triggered a `badgeOnly` notif, and made sure that both clients did not display a system notif when backgrounded but did display an in-app notif when foregrounded.

Reviewers: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D309