[lib] threadIsGroupChat
Summary:
I noticed that the name of `threadIsPersonalChat` was now confusing, since D294 introduces `threadTypes.PERSONAL`. I figured I should rename it.

I looked at the callsites and realized that we could combine it with `threadIsTwoPersonChat`. In the `MessagePreview` callsites, my new function should be logically equivalent to the old condition. In the notif callsites, we can assume at least two members, so it's equivalent.

I also think it makes sense to combine these in light of the work @KatPo is doing to handle permissions for blocked users. Our "condition 2" was revised today to consider both 1-person and 2-person threads, so our new `threadIsGroupChat` function would be convenient here.

Test Plan: Flow

Reviewers: palys-swm, KatPo

Subscribers: zrebcu411, Adrian, KatPo

Differential Revision: https://phabricator.ashoat.com/D310