[native] Always set ActivityIndicator color
Summary:
This is a follow-up to the RN0.63 upgrade. Full context is in [this issue](https://github.com/facebook/react-native/issues/29378).

Basically, there is a regression in RN0.63 for Android where if you don't set `color` for an `ActivityIndicator`, it won't display.

Test Plan: I tested both callsites in the Android emulator. Before this diff the `ActivityIndicator` was invisible, and after this diff it was visible. For the `MoreScreen` case, since it was theme-dependent I tested for both light and dark themes.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D330