Fix client_version_unsupported error on setDeviceToken
Summary:
I noticed this error while debugging an unrelated issue. It's happening because `checkClientSupported` sees the `deviceType` field and takes it as `platformDetails`. But it doesn't have a `codeVersion`, so `checkClientSupported` decides it is outdated.

This PR makes sure we send the whole `platformDetails` for newer clients.

Test Plan: Make sure `SET_DEVICE_TOKEN` succeeds on Android emulator while logged out

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D334