[native] Filter inset updates for Android 5
Summary:
Because we have `android:windowTranslucentNavigation` enabled starting at API level 21 (Android 5.0), the frame that React Native is rendering our app into does not change when the soft keyboard appears on Android.

But for some reason, `react-native-safe-area-context` is reporting that the frame does change for API level 21 (5.0) and 22 (5.1). The issue does not appear for Android 6 and higher, and the behavior is correct for Android 4.4 and lower.

I noticed while debugging some weird behavior relating to the keyboard on Android 5. I initially thought it was something weird with `KeyboardAvoidingView`, but eventually noticed that the Redux logger was showing an `UPDATE_DIMENSIONS` action every time the keyboard appeared.

Test Plan: I made sure the weird keyboard behavior was gone in both `LoggedOutModal` and `Calendar`

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D335