[client] Fix memoization of textMessageRules
Summary:
I noticed a bug where `textMessageRules` was being returned with the wrong `members` list. Turns out this was because Lodash's `memoize` only memoizes based on the first parameter.

This got me thinking about the ideal memoization behavior we wanted. I realized that globally memoizing on the `members` list was a bit of a memory leak.

This diff introduces a new approach. The rules get memoized on the `MessageList` level, with a max of 2 cached values for each `MessageList`, one for dark and one for light. We cache it by storing the state in React Context.

Whenever the `members` array gets recalculated, we recalculate our `textMessageRules` and throw out the old values.

Test Plan: Make sure the bug is gone: switching between threads results in recalculation of the markdown rules.

Reviewers: palys-swm, KatPo

Reviewed By: palys-swm, KatPo

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D349