[server] Guarantee that Redis messages resolve in the order they are received
Summary:
I noticed a bug where if user A creates a subthread with user B while user B has an active session, user B is shown the new subthread with no messages.

What was happening is that the `redisMessageTypes.NEW_MESSAGES` was getting delivered to the client before the `redisMessageTypes.NEW_UPDATES` with the new thread. This happened because `redisMessageTypes.NEW_UPDATES` needs to do some data-fetching before it can be resolved, whereas `redisMessageTypes.NEW_MESSAGES` gets resolved immediately.

This diff makes sure that we call `Socket.sendMessage` in the same order that the Redis messages were received.

Test Plan: Test the aforementioned bug and make sure it stops occurring

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D355