[server] Block createUpdates on redisPublish
Summary:
In order to guarantee that messages get delivered after a thread is created, I want to make sure that when `createThread`'s call to `commitMembershipChangeset` concludes, the Redis messages after the new thread have been propagated.

This doesn't actually solve any bugs – in our case, things were resolving in the right order anyways. This diff just makes sure it's guaranteed in the future.

Test Plan: Since it's not really fixing anything, not much to test. Make sure Flow is good and that things still work

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D356