[lib] Add seeMore item to chatListDataWithNestedSidebars
Summary:
We want to show a button that will launch a modal to see the rest of the sidebars if we can't show all of them.

This diff just handles the data, it doesn't actually render the see more button.

Test Plan: Make sure the sidebars still show correctly. Nothing actually changes visually

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D358