[server] Update badge counts on mark-as-unread
Summary: I noticed an issue where I marked a thread as unread on the web client and the badge count on my phone didn't update. After reading through the code, I realized we never actually considered this scenario.

Test Plan: Mark a thread as unread on the web client for a user that has a `device_token` in the DB. Inspect the `notifications` rows that get created in the database. Also tests notif sends/rescinds

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D363