[server] Create new thread when friendship request is sent
Summary: Every time the status is changed to FRIEND create a new thread. This behavior will be changed later so that at most one PERSONAL thread will exist.

Test Plan: I verified that sending a friend request results in a creation of a new thread. It works both in the prompt and in the more tab (even for couple of users). Canceling the request does not create a thread. I also verified that the error handling is not affected: sent a request with invalid user id and observed error message.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D369