[server] Call updateBadgeCount for mark-as-read
Summary:
In D363, I created a new function `updateBadgeCount` to handle updating client badge count for mark-as-unread. I didn't bother for mark-as-read because I thought `rescindPushNotifs` would handle it.

But that's not always the case. In particular, if a thread becomes unread by being marked as unread, there won't be any push notifs to rescind when it is marked as read, since only `new_message` notifs are rescindable.

This diff makes sure we always run `updateBadgeCount` for mark-as-read. I make sure we only send a badge update notif to device tokens that aren't receiving a rescind from `rescindPushNotifs`.

Test Plan: Make sure rows get added to the `notifications` table on mark-as-read, and make sure the rows look good. It's hard to actually test notif delivery in our dev environment unfortunately...

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D366