[server] Don't take threadID in updateBadgeCount
Summary:
The next diff is going to make it so `updateBadgeCount` gets called by `activityUpdater`. But unlike `setThreadUnreadStatus`, `activityUpdater` isn't necessarily dealing with a single `threadID`.

We've only been taking `threadID` in `updateBadgeCount` for logging purposes. We could update it to take a list of IDs, but honestly I don't think the value is high enough for the complexity it would require.

Note that in order to make it so `updateBadgeCount` doesn't need a `threadID` parameter, we also need to make the `thread` column in the `notifications` table optional.

Test Plan:
1. Run the script
2. Make sure notif sends don't cause any errors
3. Make sure mark-as-read and mark-as-unread don't cause any errors
4. Make sure `rescindPushNotifs` doesn't cause any errors

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D367