[server] Make sure activityUpdater calls updateBadgeCount
Summary:
After D363, it's now possible for a thread is be marked as unread in the database without any corresponding rescindable notifs. This occurs when a thread is marked as unread.

This diff makes sure that if a thread is marked as unread, and then the user opens up that thread, we still correctly set all of their devices' badge counts, even though there is no notif to rescind for those devices.

Test Plan:
1. Make sure rows get added to the `notifications` table when I navigate to a thread on the web client after marking it as unread.
2. Make sure the rows look good. It's hard to actually test notif delivery in our dev environment unfortunately...

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D368