Remove visibilityRules from RawThreadInfo
Summary:
`visibilityRules` is the old name for `type`, which I changed in `codeVersion` 9. I kept it around in `RawThreadInfo` because I wanted to continue supporting older clients at the time.

We [don't support](https://github.com/Ashoat/squadcal/commit/bdc6076f9d519b794a2f56223354870ad0bfc718) `codeVersion` below 24 anymore, so we can get rid of this. But we have to consider the effects on current clients' `threadStore`. If we change it on the server now, there will be a lot of inconsistency reports that get generated due to state check seeing `visibilityRules` on the client, but not on the server.

This diff avoids creating inconsistency reports. The server will start delivering `RawThreadInfo`s without `visibilityRules` as of `codeVersion` 70, and will fix the `threadStore` as of `stateVersion` 19, which will be incremented at the same time as `codeVersion`.

Test Plan:
1. I ran the code on an iOS simulator
2. I confirmed `visibilityRules` was still there using the Redux panel of the React Native Debugger
3. I waited for a state check to make sure it didn't detect any issues
4. I then incremented both `codeVersion` and `stateVersion` and reran the app on the simulator
5. I confirmed `visibilityRules` was now gone
6. I waited for a state check to make sure it didn't detect any issues

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D386