[server] Make sure targeted UPDATE_THREAD doesn't filter untargeted one
Summary:
I realized we forgot to consider something in D359. There's logic in `createUpdates` to filter updates that are handled by other updates. This logic both filters the array of `updateDatas` and deletes existing rows in the `updates` table.

With the way the code was set up, any `UPDATE_THREAD` for a given (user, thread) pair would filter/delete all previous `UPDATE_THREAD`s. But since we weren't looking at the new `targetSession` field, it was possible for an `UPDATE_THREAD` with a `targetSession` to cause other `UPDATE_THREAD`s that don't have `targetSession` to be deleted.

The existing mechanism we had to address this was to include the target in the `conditionKey`. But this only works if an `updateType` always has a target. As a result I had to think of a new mechanism.

Once I started reading through the code in `createUpdates` I got confused pretty fast. I decided to refactor it in a more readable way.

Test Plan: Generate some updates on my local instance and make sure they look good

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D398