[server] Set permissions of personal thread
Summary:
This is a minimal reasonable set of permissions.
I also considered adding `EDIT_THREAD`, but it allows changing a name, which I'm not sure if we want.
We also need to decide if we want `CREATE_SUBTHREADS` and `CREATE_SIDEBARS`.

Test Plan: I created a new personal thread and verified that I can send a message, but can't change members

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, zrebcu411, KatPo

Differential Revision: https://phabricator.ashoat.com/D412