Make sure lint-staged checks all dependent projects
Summary:
I noticed that `lint-staged` failed to catch an issue in D421 where an export was removed from `lib` that `server` depended on.

Running Flow in `server` would've caught it, but `lint-staged` is only set up to do that for changes in the `server` folder.

This diff updates `lint-staged` so that it runs Flow in all directories whenever something in `lib` changes. It also makes sure to run Flow in `server` when something in `web` changes.

Test Plan: Remove `threadIsWithBlockedUserOnly` from export list in `lib/shared/thread-utils.js` and make sure `lint-staged` errors

Reviewers: KatPo, palys-swm

Reviewed By: KatPo

Subscribers: zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D425