[lib] Exclude anonymous users in userInfoSelectorForPotentialMembers
Summary:
I noticed that opening `ComposeThread` in my dev environment on master was crashing the app. Turns out this was actually broken in D172, when we took out the `.username` check. But it only surfaced after D405, since before that, the anonymous users were being excluded for other reasons.

Flow ideally would've caught this, since without that check a `UserInfo` can't be considered an `AccountUserInfo`.

Test Plan: Make sure the crash is gone

Reviewers: KatPo, palys-swm

Reviewed By: KatPo

Subscribers: zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D428