[lib] Make user types $ReadOnly
Summary:
I was investigating why Flow failed to catch the issue introduced in D172 and fixed in D428.

It turns out that defining `AccountUserInfo` as `{| ... |} & UserInfo` was confusing Flow. I initially did it this way because for some reason Flow didn't think `AccountUserInfo` counted as an `UserInfo`, even though it looked like just a more specific version of `UserInfo`. Adding the `{| ... |} & UserInfo` hack got rid of the error.

At the time I didn't realize it, but it turns out that Flow error was entirely correct: `AccountUserInfo` is not a `UserInfo`. It's because if you pass an `AccountUserInfo` to a function that takes `UserInfo`, that function could then set `UserInfo.username` to `null`, which would break the caller's type. Flow docs have some [more details](https://flow.org/en/docs/lang/variance/) about this concept, called "Type Variance".

This is a great example of why `$ReadOnly` is important. By making all these types `$ReadOnly` we make it so `AccountUserInfo` is indeed always a `UserInfo`, since no function that takes `UserInfo` would be able to override the `username` field. That lets us get rid of the `{| ... |} & UserInfo` hack, which was breaking Flow's ability to correctly typecheck for some reason.

Test Plan:
1. Flow
2. Remove the runtime check on `username` added in D428 and make sure it causes a Flow error

Reviewers: KatPo, palys-swm

Reviewed By: KatPo

Subscribers: zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D430